7 rename db connection file
