+++ server/test/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 from starlette.testclient import TestClient
 
-from database.master_connection import ConnectionManager
+from database.database_connection import ConnectionManager
 from api.routes import application
